requirements.txt
pytest>=7.0
python-dotenv>=1.0

# Optional manual-demo extras
# Install local embedder support with:
#   pip install sentence-transformers
# Install OpenAI embedder support with:
#   pip install openai
# Install both optional embedders with:
#   pip install openai sentence-transformers
# sentence-transformers>=3.0
# openai>=1.0

# Optional: for bonus tasks
# chromadb>=0.4.0      # Real vector database integration

# Crawler and Processor dependencies
aiohttp>=3.9.0
python-docx>=1.1.0
tqdm>=4.66.0
rank-bm25>=0.2.2
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
weaviate-client>=4.0.0
packaging>=23.0
